Fixed issue where if the two tapes closest to the center were not valid target nothing would be locked on to
Fixed issue where if the two tapes closest to the center were not valid targets then nothing would be locked on to. Now the pair of tapes that are closest to the contourDetectionX that form a valid target will be correctly locked on to if the 2 closest tapes don't form a valid target. Nothing will be locked on to if there are no valid targets.